requirements.txt
streamlit==1.50.0
pillow==11.3.0
torch==2.9.0
torchvision==0.24.0
